backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-multipart
google-api-python-client
google-auth-httplib2